refactor native functions to own package
