+++ internal/vm/natives/testdata/charts/foobar/Chart.yaml
