+++ internal/vm/natives/testdata/charts/foobar/templates/config-map.yaml
