+++ internal/vm/natives/testdata/charts/foobar/templates/secret.yaml
