+++ internal/vm/natives/testdata/charts/foobar/values.yaml
